Pulling code for changes

+++ apps/wp_importer/utils.py
@@ -19,12 +19,20 @@ from django.contrib.auth.models import User
 from django.conf import settings
 from django.template import RequestContext
 
+# Replaces shortcodes in the body of an article with appropriate HTML structures.
+  
 def replace_short_code(body):
+
     body = re.sub("(.*)(\\[caption.*caption=\")(.*)(\"\\])(.*)(<img.*(\"|/| )>)(.*)(\\[/caption\\])(.*)", "\\1\\6<div class=\"caption\">\\3</div>\\10", body)      
     body = re.sub("(.*)(\\[gallery?.*?\\])(.*)", '', body)
     return body
 
 def get_posts(items, uri_parser, user):
+    """
+    Find each item marked "post" in items
+    If that Article has already been created, skip it.
+    If not, create Article object and Redirect object.
+    """
     post_list = []
     redirect_list = []
     alreadyThere = False
@@ -90,8 +98,11 @@ def get_posts(items, uri_parser, user):
                 })
 
     return post_list, redirect_list
+
+# Find each item marked "page" in items.If that Page has already been created, do nothing. If not, create Page object.
     
 def get_pages(items, uri_parser, user):
+ 
     page_list = []
     alreadyThere = False
 
@@ -139,37 +150,64 @@ def get_pages(items, uri_parser, user):
 
     return page_list
 
+#   Find any URL contained in an "attachment" If that File has already been created, skip it. If not, go to the URL, and save the media there as a File.
+  
 def get_media(items, uri_parser, user):
+    
     for node in items:
         post_type = node.find('wp:post_type').string
         if post_type == 'attachment':
-            media_url = node.find('wp:attachment_url').string
-            source = urllib2.urlopen(media_url).read()
-            media_url = uri_parser.parse(media_url).file
+            media_url_in_attachment = node.find('wp:attachment_url').string
+            media_url = uri_parser.parse(media_url_in_attachment).file
             media_url = os.path.join(settings.MEDIA_ROOT, media_url)
 
-            with open(media_url, 'wb') as f:
-                f.write(source)
-                file_path = f.name
+            alreadyThere = False
 
-            new_media = File(guid=unicode(uuid.uuid1()), file=file_path, creator=user, owner=user)
-            new_media.save()
+            for media_path in File.objects.all():
+                if media_url == media_path.file:
+                    alreadyThere = True
+                    break
+            
+            if not alreadyThere:
+                source = urllib2.urlopen(media_url_in_attachment).read()
 
+                with open(media_url, 'wb') as f:
+                    f.write(source)
+                    file_path = f.name
+
+                new_media = File(guid=unicode(uuid.uuid1()), file=file_path, creator=user, owner=user)
+                new_media.save()
+
+# For each File in the database, replace an instance of that file's URL in the given HTML with a file path.
+    
 def correct_media_file_path(body, uri_parser):
-    for url in File.objects.all():
-        media_file = unicode(url.file)
-        match = re.search("(.*)(http://.*\\/?\\/\\b\\S+\\/)(\\w.*?)(\\\".*)", body)
+    
+    for overwrite in File.objects.all():
+        print 'Got to the for loop'
+        print overwrite.basename() + '\n'
+        match = re.search("(.*)(http://.*\\/?\\/\\b\\w+\\/)((\\S+)(\\-\\d+x.*?\\S*.*)|(\\S+.*?\\S+))(\\.\\S+)(\\\".*)", body)
         if match:
             match.group()
-            if media_file.endswith(match.group(3)):
-                body = re.sub("(.*)(http://.*\\/?\\/\\b\\S+\\/)(" + re.escape(match.group(3)) + ".*?)(\\\".*)", "\\1/site_media/media/" + match.group(3) + "/\\4", body)
-                # above: site_media/media (with match.group(3)) is hardcoded 
-                # rather than media_file because the resulting url from 
-                # media_file (with settings.MEDIA_ROOT) breaks - possible fix??
+            if overwrite.basename() == match.group(3) + match.group(7):
+                print 'I am in the if statement before sub'
+                print overwrite.basename() + '\n'
+                #media_overwrite.endswith(match.group(3)+match.group(7)):
+                #body = re.sub("(.*)(http://.*\\/?\\/\\b\\w+\\/)(" + re.escape(overwrite.basename())")(\\.\\S+)(\\\".*)", "\\1/files/" + match.group(3) + "/\\7", body)
+                body = re.sub("(.*)(http://.*\\/?\\/\\b\\w+\\/)(" + re.escape(overwrite.basename()) + ".*?)(\\\".*)", "\\1/files/" + str(overwrite.pk) + "/\\4", body)
+                print 'I am after re.sub'
+                print overwrite.basename() + '\n'
 
+            elif overwrite.basename() == match.group(3) + match.group(7):
+                print 'In the elif before re.sub'
+                body = re.sub("(.*)(http://.*\\/?\\/\\b\\w+\\/)("+ re.escape(match.group(6)) +")(\\.\\S+)(\\\".*)", "\\1/files/" + str(overwrite.pk) + "/\\7", body)
+                print 'In elif after re.sub'
     return body
+    print 'After return'
+
+# Parse the given xml file using BeautifulSoup. Save all Article, Redirect and Page objects.
 
 def run(file_name, request):
+    
     f = open(file_name, 'r')
     xml = f.read()
     f.close()
